Cleanup up navigation code so that it's separated into its components.
